- jbj: eliminate rpmal debugging code.

+++ lib/rpmgi.c
@@ -211,6 +211,7 @@ static rpmRC rpmgiLoadReadHeader(rpmgi gi)
 	/* Not a header, so try for a manifest. */
 	gi->argv[gi->i] = NULL;		/* Mark the insertion point */
 	rpmrc = rpmgiLoadManifest(gi, fn);
+	/* XXX its unclear if RPMRC_NOTFOUND should fail or continue here. */
 	if (rpmrc != RPMRC_OK) {
 	    gi->argv[gi->i] = fn;	/* Manifest failed, restore fn */
 	    break;
@@ -252,7 +253,7 @@ rpmlog(RPMLOG_DEBUG, "FTS_%s\t%*s %s%s\n", ftsInfoStr(fts->fts_info),
     case FTS_DP:	/* postorder directory */
 	break;
     case FTS_F:		/* regular file */
-	if (fts->fts_namelen <= sizeof(".rpm"))
+	if ((size_t)fts->fts_namelen <= sizeof(".rpm"))
 	    break;
 	/* Ignore all but *.rpm files. */
 	s = fts->fts_name + fts->fts_namelen + 1 - sizeof(".rpm");
